backend/app/agents/gemini_agent.py
"""
LEGACY GEMINI AGENTS - REPLACED BY PORTIA INTEGRATION

These agents are kept for reference but the new Portia-powered agents
in portia_agents.py provide MAXIMUM integration with:
- Structured planning
- Tool orchestration
- Human-in-the-loop workflows
- Cloud storage
- Multi-tool coordination

For production use, use the agents from portia_agents.py instead.
"""

import google.generativeai as genai
from typing import Dict, Any, Optional
import json
import asyncio
from ..config import settings

class GeminiAgent:
    """Base class for Google Gemini AI agents"""
    
    def __init__(self, name: str, description: str, instructions: str):
        self.name = name
        self.description = description
        self.instructions = instructions
        
        # Configure Gemini
        genai.configure(api_key=settings.GOOGLE_GEMINI_API_KEY)
        self.model = genai.GenerativeModel('gemini-pro')
        
    async def run(self, **kwargs) -> Dict[str, Any]:
        """Run the agent with given parameters"""
        try:
            # Build the prompt
            prompt = self._build_prompt(**kwargs)
            
            # Generate response
            response = await self._generate_response(prompt)
            
            return {
                "status": "success",
                "agent": self.name,
                "result": response,
                "metadata": {
                    "model": "gemini-pro",
                    "description": self.description
                }
            }
            
        except Exception as e:
            return {
                "status": "error",
                "agent": self.name,
                "error": str(e),
                "metadata": {
                    "model": "gemini-pro",
                    "description": self.description
                }
            }
    
    def _build_prompt(self, **kwargs) -> str:
        """Build the prompt for the agent"""
        prompt = f"""
You are {self.name}: {self.description}

Instructions: {self.instructions}

Please process the following request and provide a detailed, structured response:

"""
        
        # Add input parameters
        for key, value in kwargs.items():
            prompt += f"{key}: {value}\n"
        
        prompt += "\nPlease provide your response in a structured format."
        return prompt
    
    async def _generate_response(self, prompt: str) -> str:
        """Generate response from Gemini"""
        try:
            response = self.model.generate_content(prompt)
            return response.text
        except Exception as e:
            return f"Error generating response: {str(e)}"

class JobTrackerAgent(GeminiAgent):
    """Agent for tracking job applications and syncing to Google Drive"""
    
    def __init__(self):
        super().__init__(
            name="Job Application Tracker Agent",
            description="Tracks job applications, responses, and syncs data to Google Drive",
            instructions="""
            - Track job application status (applied, interview, offer, rejected)
            - Monitor response times and follow-up actions
            - Sync data to Google Sheets/Excel
            - Provide insights on application patterns
            - Suggest follow-up actions
            """
        )

class JobApplicationAgent(GeminiAgent):
    """Agent for auto-applying to jobs and sending outreach"""
    
    def __init__(self):
        super().__init__(
            name="Job Application Agent", 
            description="Automatically applies to jobs and sends personalized outreach",
            instructions="""
            - Read resume and job requirements
            - Generate personalized cover letters
            - Auto-apply to matching jobs
            - Send cold DMs/emails to hiring managers
            - Track application responses
            - Optimize application strategy
            """
        )

class CareerRoadmapAgent(GeminiAgent):
    """Agent for generating personalized career roadmaps"""
    
    def __init__(self):
        super().__init__(
            name="Career Roadmap Agent",
            description="Creates personalized career development plans",
            instructions="""
            - Analyze current skills and experience
            - Identify career goals and timeline
            - Generate skill development roadmap
            - Suggest learning resources (courses, books, projects)
            - Track progress on Leetcode, GitHub, LinkedIn
            - Provide milestone checkpoints
            """
        )

class AnalyticsAgent(GeminiAgent):
    """Agent for generating analytics and scorecards"""
    
    def __init__(self):
        super().__init__(
            name="Analytics Agent",
            description="Generates insights and visualizations for career progress",
            instructions="""
            - Analyze application funnel metrics
            - Track GitHub contribution patterns
            - Monitor Leetcode rating progress
            - Measure LinkedIn engagement
            - Generate candidate scorecards
            - Provide actionable insights
            - Create progress reports
            """
        )

class NetworkingAgent(GeminiAgent):
    """Agent for LinkedIn networking and connections"""
    
    def __init__(self):
        super().__init__(
            name="Networking Agent",
            description="Finds similar profiles and manages LinkedIn connections",
            instructions="""
            - Find professionals with similar backgrounds
            - Generate personalized connection messages
            - Send LinkedIn connection requests
            - Track connection acceptance rates
            - Suggest networking events and groups
            - Maintain professional relationships
            """
        )

class CandidateIntakeAgent(GeminiAgent):
    """Agent for processing candidate applications"""
    
    def __init__(self):
        super().__init__(
            name="Candidate Intake Agent",
            description="Processes candidate applications and generates intake forms",
            instructions="""
            - Generate Google Forms for job applications
            - Collect and organize candidate data
            - Screen applications for basic requirements
            - Categorize candidates by role and experience
            - Sync data to recruitment dashboard
            - Generate candidate summaries
            """
        )

class CandidateRankingAgent(GeminiAgent):
    """Agent for ranking and scoring candidates"""
    
    def __init__(self):
        super().__init__(
            name="Candidate Ranking Agent",
            description="Analyzes and ranks candidates based on multiple criteria",
            instructions="""
            - Analyze resumes for relevant experience
            - Evaluate GitHub activity and code quality
            - Assess Leetcode problem-solving skills
            - Review LinkedIn professional presence
            - Score candidates on technical and soft skills
            - Generate ranking reports
            - Provide hiring recommendations
            """
        )

class InterviewSchedulingAgent(GeminiAgent):
    """Agent for scheduling interviews and generating questions"""
    
    def __init__(self):
        super().__init__(
            name="Interview Scheduling Agent",
            description="Schedules interviews and generates personalized questions",
            instructions="""
            - Integrate with Google Calendar for scheduling
            - Auto-schedule interviews based on availability
            - Generate personalized interview questions
            - Prepare candidate-specific assessments
            - Send interview confirmations and reminders
            - Track interview outcomes
            - Provide interview feedback templates
            """
        )
